Move AtHop to serialonly
