+++ src/main/java/au/id/micolous/metrodroid/card/desfire/DesfireCard.java
@@ -46,7 +46,7 @@ import au.id.micolous.metrodroid.transit.serialonly.MykiTransitData;
 import au.id.micolous.metrodroid.transit.opal.OpalTransitData;
 import au.id.micolous.metrodroid.transit.orca.OrcaTransitData;
 import au.id.micolous.metrodroid.transit.adelaide.AdelaideMetrocardTransitData;
-import au.id.micolous.metrodroid.transit.stub.AtHopStubTransitData;
+import au.id.micolous.metrodroid.transit.serialonly.AtHopStubTransitData;
 import au.id.micolous.metrodroid.transit.tfi_leap.LeapTransitData;
 import au.id.micolous.metrodroid.transit.tfi_leap.LeapUnlocker;
 import au.id.micolous.metrodroid.transit.serialonly.TrimetHopTransitData;
@@ -65,7 +65,6 @@ import java.security.AccessControlException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
-import java.util.Collection;
 import java.util.GregorianCalendar;
 import java.util.List;
 import java.util.Locale;
